edits

+++ SBL/SB2_ParameterSettings.py
@@ -61,6 +61,7 @@
 #
 # Contact the author: m a i l [at] m i k e t i p p i n g . c o m
 #
+import numpy as np
 
 def SB2_ParameterSettings(*args):
     
@@ -72,7 +73,7 @@ def SB2_ParameterSettings(*args):
     
     # Any settings specified?
     
-    numSettings    = len(args)/2
+    numSettings    = np.int(len(args)/2)
     
     ## Defaults - over-ridden later if requested
